+++ gz/rules_test4_example_com.php
@@ -1,4 +1,10 @@
-<?php return array (
+<?php
+/**
+ * test4.example.com 域名广告和插播规则
+ * 清理于: 2026-07-07 22:40:05
+ */
+
+return array (
   'domain' => 'test4.example.com',
   'duration_rules' => 'bad',
   'discontinuity_rules' => 'bad',